requirements.txt
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.2
torch>=2.0.0
fastapi>=0.103.1
uvicorn>=0.23.2
requests>=2.31.0 